Add empty error listener to client
This is being done to avoid crashing the Node.js process for cases where
the client creator does not register its own listener. Errors shouldn't
necessarily cause a crash since they can be ephemeral - especially when
a list of brokers is provided in the config and a failover reconnect to
another broker can be performed.