Rework subscription tracking in subs_count_test
Previously, the subs_count_test waited for the subacks received by the
client to equal the number of subscriptions sent before testing the
subscription count with the broker. subacks for internal topics that the
client could subscribe for, though, could cause the test to ask the
broker for the number of subscriptions prematurely. This commit reworks
the logic to ensure that all of the subacks for the topics that the test
is specifically subscribing for have been received before testing the
subscription count with the broker.