Handle resubscription logic in DXL JavaScript client directly
Previously, the logic for handling resubscriptions across connections
was deferred to the MQTT.js library. The MQTT.js library queues up
subscribe / unsubscribe packets each time the corresponding client
function is invoked. If the sub/unsub packet is processed while
the client is disconnected, the packet could end up being dropped/not
acked. If the packet is processed while the client is in the process of
disconnecting, the client could end up not being shutdown - due to an
outstanding packet that is never acked.

This commit disables the use of resubscribe in the MQTT.js library,
instead issuing subscribe calls to the broker for active subscriptions
whenever a reconnect occurs. Unsubscribe packets are only attempted when
the client is connected, preventing the stalled shutdown issue.