Coerce message payload to string when not a Buffer or String
Previously, if the message payload had not been set to either a Buffer
or String type, it could be malformed when serialized and sent through a
broker. For this commit, any non-Buffer type value for a message payload
is coerced into a String type value prior to serialization. This allows
a binary payload to be preserved when sent through a broker.