Add support for finding openssl and configuration used by CLI
This commit adds support to the CLI subcommands for finding the openssl
executable to be spawned. This includes an option for specifying the
full path on the command line, --opensslbin, and falling back on Windows
to the default installation directories for the slproweb OpenSSL package
when openssl cannot be found in the system path.

The commit also autogenerates a mostly empty openssl.cnf file to be used
with the openssl commands when the default one which openssl would use
does not exist.

This commit pulls in two new dependencies, tmp and which. These
dependencies are used for creating a temporary openssl.cnf file and for
attempting to find openssl in the system path.