Change `requireClient` function to `require` in common.js
This commit changes the `requireClient` function in `common.js` to
be named `require`. The implementation of `require` now handles
the ability to require the SDK package by explicit name when
running either from source or a package install. The implementation
also adjusts the module path if needed in order to resolve a
dependency installed for various NPM versions.