Populate 'code' property in MessageError with string value
Previously, the 'code' property for a MessageError contained the
errorCode value, a numeric type, from the DXL ErrorResponse. With
Node.js 8+, the 'code' property for derivatives of the Error class
(like MessageError) is expected to include a string-type value rather
than a numeric one. For compatibility with the newer Node.js versions,
this commit now populates a string-based value in the MessageError.code
property. The code property is only set to a non-empty string for
'well-defined' broker errors. For example, the value would be populated
with 'ERR_DXL_SERVICE_UNAVAILABLE' when the broker returns an
ErrorResponse for a request that cannot be routed.

The numeric DXL error code property is now available as a separate
top-level property in the MessageError class, 'dxlErrorCode'. Also, an
alias was created for the 'detail' property in the MessageError class,
'dxlMessage'. These 'dxl' prefixed property names allow for consumers
receiving a MessageError to more easily use duck typing to derive the
DXL-level error details - avoiding the need to use the instanceof to
distinguish a MessageError from other Error-type classes. instanceof can
be unreliable in cases where a MessageError is being evaluated across
packages which import separate instances of the
opendxl-client-javascript library from within the same node process.