Change MessageError to RequestError and use string for code property
This commit renames the MessageError class to RequestError in order
to tie it more directly to its use in the response to a client
asyncRequest().

Previously, the 'code' property for a MessageError contained the
errorCode value, a numeric type, from the DXL ErrorResponse. With
Node.js 8+, the 'code' property for derivatives of the Error class
(like MessageError) is expected to include a string-type value rather
than a numeric one. For compatibility with the newer Node.js versions,
this commit now populates a string-based value in the RequestError.code
property. The code property is only set to a non-empty string for
'well-defined' broker errors. For example, the value would be populated
with 'ERR_DXL_SERVICE_UNAVAILABLE' when the broker returns an
ErrorResponse for a request that cannot be routed.

An alias was created for the 'detail' property in the RequestError
class, 'dxlErrorResponse'. The 'dxl' prefixed property name allows for
consumers receiving a RequestError to more easily use duck typing to
derive the DXL-level error details - avoiding the need to use the
instanceof to distinguish a RequestError from other Error-type
classes. instanceof can be unreliable in cases where a RequestError
is being evaluated across packages which import separate instances
of the opendxl-client-javascript library from within the same Node.js
process.