Add provisionConfig static function to Config class
This commit adds a provisionConfig static function to the `Config`
class, to be used for non-CLI based provisioning of a client
configuration. This commit also allows for the `san` (subject
alternative name) option for CSR generation and provisioning functions
to be omitted. For use in testing the new `provisionConfig` function,
the CLI test helpers related to PKI functionality were separated into a
new module for reuse.