Write errors to log when no callback is supplied to PKI function
Previously, if an error occurred during invocation of a PKI-related
function but no callback was provided to the function, the error would
essentially just be swallowed. With the changes in this commit, the
error would be written to the log when no callback is provided.

If a callback is provided, the error is not automatically logged.
This avoids duplicating error logging that the receiver of the
callback would do.